add condition if there is a user (security) / if not, redirect to login

+++ src/Controller/ProgrammeController.php
@@ -17,9 +17,17 @@ class ProgrammeController extends AbstractController
     #[Route('/programme', name: 'app_programme')]
     public function index(): Response
     {
-        return $this->render('programme/index.html.twig', [
-            'controller_name' => 'ProgrammeController',
-        ]);
+
+        //On vérifie s'il y a un user (comme ça pas de modif possible autrement)
+        if($this->getUser()) {
+            
+            return $this->render('programme/index.html.twig', [
+                'controller_name' => 'ProgrammeController',
+            ]);
+
+        } else {
+            return $this->redirectToRoute("app_login");
+        }
     }
 
     #[Route('/programme/add/{idSession}', name: 'add_programme')]
@@ -27,32 +35,32 @@ class ProgrammeController extends AbstractController
     {
 
         //On vérifie s'il y a un user (comme ça pas de modif possible autrement)
-        // if($this->getUser()) {
+        if($this->getUser()) {
             
-        // } else {
-        //     return $this->redirectToRoute("app_login");
-        // }
+            $form = $this->createForm(ProgrammeType::class, $programme);
+            $form->handleRequest($request);
 
-        $form = $this->createForm(ProgrammeType::class, $programme);
-        $form->handleRequest($request);
+            if($form->isSubmitted() && $form->isValid()) {
+                $programme = $form->getData();
+                $entityManager = $doctrine->getManager();
+                $session = $entityManager->getRepository(Session::class)->find($idSession);  
+                /* On utilise la méthode créée de base dans Session grâce au ManyToMany */
+                $session->addProgramme($programme);
+                $entityManager->persist($programme);
+                // $entityManager->persist($programme);
+                $entityManager->flush();
 
-        if($form->isSubmitted() && $form->isValid()) {
-            $programme = $form->getData();
-            $entityManager = $doctrine->getManager();
-            $session = $entityManager->getRepository(Session::class)->find($idSession);  
-            /* On utilise la méthode créée de base dans Session grâce au ManyToMany */
-            $session->addProgramme($programme);
-            $entityManager->persist($programme);
-            // $entityManager->persist($programme);
-            $entityManager->flush();
+                return $this->redirectToRoute('show_session',
+                ['id' => $session->getId()]);
+            }
 
-            return $this->redirectToRoute('show_session',
-            ['id' => $session->getId()]);
+            return $this->render('session/show.html.twig', [
+                'formAddProgramme' => $form->createView() 
+            ]);
+
+        } else {
+            return $this->redirectToRoute("app_login");
         }
-
-        return $this->render('session/show.html.twig', [
-            'formAddProgramme' => $form->createView() 
-        ]);
     }
 
     #[Route('/programme/delete/{id}/{idSession}', name: 'delete_programme')]
@@ -60,22 +68,21 @@ class ProgrammeController extends AbstractController
     {
 
         //On vérifie s'il y a un user (comme ça pas de modif possible autrement)
-        // if($this->getUser()) {
-            
-        // } else {
-        //     return $this->redirectToRoute("app_login");
-        // }
+        if($this->getUser()) {
+           
+            /* Ici, on fait appel à deux fonctions présentes de base dans le ProgrammeRepository : find() et remove() */
+            $entityManager = $doctrine->getManager();
+            $programme = $entityManager->getRepository(Programme::class)->find($id);
+            $entityManager->remove($programme);
+            /* flush() sauvegarde les changements effectués en base de données */
+            $entityManager->flush();
 
-        /* Ici, on fait appel à deux fonctions présentes de base dans le ProgrammeRepository : find() et remove() */
-        $entityManager = $doctrine->getManager();
-        $programme = $entityManager->getRepository(Programme::class)->find($id);
-        $entityManager->remove($programme);
-        /* flush() sauvegarde les changements effectués en base de données */
-        $entityManager->flush();
+            return $this->redirectToRoute('show_session',
+            ['id' => $idSession]);            
 
-        return $this->redirectToRoute('show_session',
-    ['id' => $idSession]);
-
+        } else {
+            return $this->redirectToRoute("app_login");
+        }
     }
